// src/app/models/demo.inputs.ts
// --- Generated by Signaloid CLI ---

// Type definitions for the inputs that will be used in the Angular application.
export type InputType = 'text' | 'basic-slider' | 'distribution-slider' | 'multiple-choice';

export interface BaseInputConfig {
  id: string;
  name: string;
  argumentFlag: string;
  type: InputType;
}

export interface TextInputConfig extends BaseInputConfig {
  type: 'text';
  initialValue: string;
}

export interface SliderConfig extends BaseInputConfig {
  type: 'basic-slider' | 'distribution-slider';
  min: number;
  max: number;
  step: number;
  initialValue: number;
  xAxisLabel: string;
  adjustWidth: boolean;
  isDistribution: boolean;
  initialDistribution?: [number, number][];
}

export interface MultipleChoiceConfig extends BaseInputConfig {
  type: 'multiple-choice';
  options: { name: string; value: string }[];
  initialValue: string;
}

export type InputConfig = TextInputConfig | SliderConfig | MultipleChoiceConfig;

// The array of configured inputs, generated based on user prompts.
export const demoInputs: InputConfig[] = [
  {
    id: 'temperature',
    name: 'Temperature (ADC Reading)',
    min: 250000,
    max: 700000,
    step: 25000,
    initialValue: 525000,
    argumentFlag: ' -t ',
    type: 'distribution-slider',
    xAxisLabel: 'Temperature (ADC Reading)',
    adjustWidth: false,
    isDistribution: true,
    initialDistribution: [
      [500000, 75],
      [525000, 75],
      [550000, 75],
    ],
  },
  {
    id: 'pressure',
    name: 'Pressure (ADC Reading)',
    min: 200000,
    max: 800000,
    step: 25000,
    argumentFlag: ' -p ',
    xAxisLabel: 'Pressure (ADC Reading)',
    adjustWidth: false,
    type: 'distribution-slider',
    isDistribution: true,
    initialValue: 300000,
    initialDistribution: [
      [275000, 75],
      [300000, 75],
      [325000, 75],
    ],
  },
  {
    id: 'humidity',
    name: 'Humidity (ADC Reading)',
    min: 12000,
    max: 30000,
    step: 1000,
    initialValue: 15000,
    initialDistribution: [
      [14000, 75],
      [15000, 75],
      [16000, 75],
    ],
    argumentFlag: ' -u ',

    xAxisLabel: 'Humidity (ADC Reading)',
    adjustWidth: false,
    isDistribution: true,
    type: 'distribution-slider',
  },
];